windows-client/src/app/services/communication-socket/communication-socket.service.ts
import { Injectable } from '@angular/core';
import { io, Socket } from 'socket.io-client';
import { environment } from 'src/environments/environment';

@Injectable({
    providedIn: 'root',
})
export class CommunicationSocketService {
    // code imported from https://gitlab.com/nikolayradoev/socket-io-exemple
    // Author: Nikolay Radoev
    private socket: Socket;

    isSocketAlive() {
        return this.socket && this.socket.connected;
    }

    connect() {
        this.socket = io(environment.serverBaseUrl, { transports: ['websocket'], upgrade: false });
        this.socket.emit('connection', 'test');
    }

    disconnect() {
        this.socket.disconnect();
    }

    on<T>(event: string, action: (data: T) => void): void {
        this.socket.on(event, action);
    }

    off(event: string): void {
        this.socket.off(event);
    }

    send<T>(event: string, data?: T): void {
        if (data) {
            this.socket.emit(event, data);
        } else {
            this.socket.emit(event);
        }
    }

    getSocketId(): string {
        return this.socket.id;
    }
}
